refactor(spdx-utils): Move operator-relared code to a separate file
Improve overview a bit my splitting the large `Extensions.kt` and
grouping operator code.
